requirements.txt
flask==1.0.2
flask-script==2.0.6
flask-sqlalchemy==2.4.0
flask-mysqldb==0.2.0
flask-wtf==0.14.2
wtf==0.1
flask-migrate==2.4.0
xmltodict==0.12.0
flask-session==0.3.1
redis==3.2.1
sqlalchemy-utils==0.33.11
sqlalchemy-mptt==0.2.4
flask_cors==3.0.7
flask_restful==0.3.7
pillow==6.0.0
ConfigParser==3.7.4
requests==2.21.0
blinker==1.4
qiniu==7.2.4
python-alipay-sdk==1.10.0
